Refactor: renamed directory from *_payload to *_response
